Add support for URG-04LX in SCIP 1.1 mode
The urg_node does not support SCIP 1.1. The Hokuyo URG-04LX supports both
SCIP 1.1 and SCIP 2.0, but needs to be switched to SCIP 2.0 at every startup
in its default configuration. For this purpose the function
URGCWrapper::setToSCIP2() was added.
A URG-04LX in SCIP 1.1 mode used to lead to an exception being thrown in
URGCwrapper::initialize. Now, before throwing the exception an attempt to
switch the sensor to SCIP 2.0 is made.