synchronize_time: reset when clock is warped
If either the hardware clock or system clock warp, reset the EMA to
prevent incorrect clock values from being used. Detect the warp by
putting a limit on the absolute error between the synchronized clock
and the system clock. When a warp is detected, reset the EMA to force
the clocks to resynchronize. Use the system clock until the EMA has
stabalized again.