Better validation of ip address for set_urg_ip
Use the ipaddress library to validate the given ip-addresses.

Don't use a loop to fill up missing zeros.

Before the given ip was checked for their format
which means that the format is like AAA.BBB.CCC.DDD.
But it was never checked if it is actually a valid ip and not only
in the correct format.

Parsing it with ipaddress will ensure that the given ip is valid.

Using zfill to fill missing zeros just feels nicer.