Prevent diagnostics segfault by accessing empty urg_ pointer (#107)
This PR prevents a potential segfault in the populateDiagnostics
function, when the driver disconnects the urg_ pointer is reset. If the
diagnostics function is in the middle of execution (it runs in a
separate thread) at the time it will try to call a function on an empty
pointer. This happened more frequently before
[PR-102](https://github.com/ros-drivers/urg_node/pull/102) but it still
could potentially happen.

The populateDiagnostics function is accessing urg_ for things that are
static or it already has access to via local variables so I've just
refactored the function so it only accesses urg_ to check if it exists.